refactor styles folder structure
