+++ components/CharacterCard/styles.module.scss
