+++ components/Episodes/styles.module.scss
@@ -1,0 +1,56 @@
+.episodesDetailsContainer {
+  display: flex;
+  flex-direction: column;
+  max-width: 620px;
+  width: 100%;
+}
+
+.episodesContainer {
+  @extend .episodesDetailsContainer;
+  border: 1px solid white;
+  border-radius: 5px;
+  padding: 10px;
+}
+
+.episodeList {
+  display: flex;
+  flex-direction: row;
+  width: 100%;
+  padding-bottom: 15px;
+  overflow-y: scroll;
+}
+
+.episodeInfo {
+  margin-bottom: 5px;
+}
+
+.episode {
+  min-width: 90px;
+  cursor: pointer;
+  background-color: transparent;
+  border: none;
+  padding: 0;
+  text-align: start;
+}
+
+.selected {
+  color: green;
+}
+
+@media (min-width: 320px) {
+  .episodeList {
+    max-width: 100%;
+  }
+}
+
+@media (min-width: 481px) {
+  .episodeList {
+    max-width: 100%;
+  }
+}
+
+@media (min-width: 641px) {
+  .episodeList {
+    max-width: 100%;
+  }
+}
