+++ components/Paginator/styles.module.scss
