+++ styles/CharacterDetails.module.scss
@@ -21,45 +21,6 @@
     "image created";
 }
 
-.episodesDetailsContainer {
-  display: flex;
-  flex-direction: column;
-  max-width: 620px;
-  width: 100%;
-}
-
-.episodesContainer {
-  @extend .episodesDetailsContainer;
-  border: 1px solid white;
-  border-radius: 5px;
-  padding: 10px;
-}
-
-.episodeList {
-  display: flex;
-  flex-direction: row;
-  width: 100%;
-  padding-bottom: 15px;
-  overflow-y: scroll;
-}
-
-.episodeInfo {
-  margin-bottom: 5px;
-}
-
-.episode {
-  min-width: 90px;
-  cursor: pointer;
-  background-color: transparent;
-  border: none;
-  padding: 0;
-  text-align: start;
-}
-
-.selected {
-  color: green;
-}
-
 .characterImage {
   grid-area: image;
 }
@@ -109,10 +70,6 @@
     padding: 60px 20px 0 20px;
   }
 
-  .episodeList {
-    max-width: 100%;
-  }
-
   .detailsGrid {
     grid-template-areas:
       "image image"
@@ -128,10 +85,6 @@
 @media (min-width: 481px) {
   .detailsContainer {
     padding: 60px 20px 0 20px;
-  }
-
-  .episodeList {
-    max-width: 100%;
   }
 
   .detailsGrid {
@@ -160,9 +113,5 @@
       "image gender"
       "image origin"
       "image created";
-  }
-
-  .episodeList {
-    max-width: 100%;
   }
 }
