+++ styles/CharacterList.module.scss
